refactor: create dictionary module
